+++ src/data/key-labels.ts
@@ -35,6 +35,15 @@ export const FN_SHIFT_KEY_LABEL: KeyLabel = {
   title: "Function Layer",
 };
 
+export const FLAG_SHIFT_KEY_LABEL: KeyLabel = {
+  type: KeyLabelType.Icon,
+  c: "counter_4",
+  layer: null,
+  shiftKey: null,
+  altGraphKey: null,
+  title: "Flag Layer",
+};
+
 export const ALT_GRAPH_KEY_LABEL: KeyLabel = {
   type: KeyLabelType.String,
   c: "ALT(GR)",
@@ -157,6 +166,16 @@ export const NON_KEY_ACTION_NAME_2_RAW_KEY_LABEL_MAP: Record<
     type: KeyLabelType.Icon,
     c: "touchpad_mouse",
     title: "Mouse Middle Click",
+  },
+  QuaternaryKeymapLeft: {
+    type: KeyLabelType.Icon,
+    c: "counter_4",
+    title: "Flag Layer (left)",
+  },
+  QuaternaryKeymapRight: {
+    type: KeyLabelType.Icon,
+    c: "counter_4",
+    title: "Flag Layer (right)",
   },
   TertiaryKeymapLeft: {
     type: KeyLabelType.Icon,