+++ src/model/device-layout.model.ts
@@ -2,18 +2,19 @@ import { Icon } from "./icon.model";
 import { Tuple } from "./tuple.type";
 
 /**
- * Shape of CharaChorder 3D device layout data. 90 keys x 3 layers = 180 action codes
+ * Shape of CharaChorder 3D device layout data. 90 keys x 3-4 layers
  */
 export interface DeviceLayout {
   id: string;
   name: string;
-  layout: Tuple<Tuple<number, 90>, 3>;
+  layout: Tuple<Tuple<number, 90>, 3> | Tuple<Tuple<number, 90>, 4>;
 }
 
 export enum Layer {
   Primary = "A1",
   Secondary = "A2" /* Numeric */,
   Tertiary = "A3" /* Functional */,
+  Quaternary = "A4" /* Flag */,
 }
 
 /*